created separate grsim communication node
